+++ lib/dealog_backoffice/messages/projectors/message_approval.ex
@@ -8,7 +8,8 @@ defmodule DealogBackoffice.Messages.Projectors.MessageApproval do
     MessageSentForApproval,
     MessageApproved,
     MessageRejected,
-    MessagePublished
+    MessagePublished,
+    MessageUpdated
   }
 
   alias DealogBackoffice.Messages.Projections.MessageForApproval
@@ -43,6 +44,10 @@ defmodule DealogBackoffice.Messages.Projectors.MessageApproval do
 
   project(%MessagePublished{} = published, fn multi ->
     Ecto.Multi.delete_all(multi, :remove_published_message, query(published.message_id))
+  end)
+
+  project(%MessageUpdated{} = updated, fn multi ->
+    Ecto.Multi.delete_all(multi, :remove_updated_message, query(updated.message_id))
   end)
 
   defp update_message_approval(multi, message_id, changes) do